adjustments to parallel processing

+++ src/wfm/base.py
@@ -1,4 +1,4 @@
-from torch import float32, set_default_dtype as torch_set_default_dtype
+from torch import float16, set_default_dtype as torch_set_default_dtype
 
 ALLOWED_LIBRARIES = [
     "numpy",
@@ -10,4 +10,4 @@ EPS = 1e-6
 UB = 1e6
 BASE_BATCH_SIZE = 512
 
-torch_set_default_dtype(float32)
+torch_set_default_dtype(float16)
